refactor: nettoyer le code des écrans d'historique et de suivi
Supprime les casts redondants dans driver_history_screen et history_screen, et retire la variable _trip inutilisée dans track_emergency_screen.

+++ lib/features/driver/driver_history_screen.dart
@@ -55,7 +55,7 @@ class _DriverHistoryScreenState extends State<DriverHistoryScreen> {
       setState(() {
         _items = data['items'] as List<dynamic>;
         _page = data['page'] as int;
-        _hasMore = (_page as int) < (data['pages'] as int);
+        _hasMore = _page < (data['pages'] as int);
       });
     } catch (_) {
       /* ignore */
@@ -76,7 +76,7 @@ class _DriverHistoryScreenState extends State<DriverHistoryScreen> {
       setState(() {
         _items = [..._items, ...(data['items'] as List<dynamic>)];
         _page = data['page'] as int;
-        _hasMore = (_page as int) < (data['pages'] as int);
+        _hasMore = _page < (data['pages'] as int);
       });
     } catch (_) {
       /* ignore */
